Todo el paquete "juego" es de clase static
