+++ src/main/java/com/dataontheroad/tute/experimento/service/ExperimentoIndividualServiceImpl.java
@@ -4,7 +4,6 @@ import com.dataontheroad.tute.juego.domain.jugador.Jugador;
 import com.dataontheroad.tute.juego.domain.partida.Partida;
 import com.dataontheroad.tute.experimento.domain.ExperimentoIndividual;
 import com.dataontheroad.tute.juego.service.partida.PartidaService;
-import com.dataontheroad.tute.juego.service.partida.PartidaServiceImpl;
 
 import java.util.List;
 
@@ -12,7 +11,7 @@ public class ExperimentoIndividualServiceImpl implements ExperimentoIndividualSe
 
     @Override
     public ExperimentoIndividual creaExperimentoIndividual(List<Jugador> jugadorList) {
-        PartidaService partidaService = new PartidaServiceImpl();
+        PartidaService partidaService = new PartidaService();
         ExperimentoIndividual experimentoIndividual = new ExperimentoIndividual();
 
         Partida partida = partidaService.crearPartida(jugadorList);
