+++ src/test/java/com/dataontheroad/tute/juego/service/partida/PartidaCierrreServiceImplTest.java
@@ -27,12 +27,12 @@ class PartidaCierrreServiceImplTest {
 
         listJugadores = creadorListaJugardoresPartidaConLaMismaEstrategia(2, (new PrimeraCartaDeLaManoStrategy()));
 
-        Partida partida = (new PartidaServiceImpl()).crearPartida(listJugadores);
+        Partida partida = (new PartidaService()).crearPartida(listJugadores);
 
         partida.getMesa().getJugadorList().get(0).setPuntuacion(50);
         partida.getMesa().getJugadorList().get(1).setPuntuacion(70);
 
-        (new PartidaServiceImpl()).cierrePartida(partida);
+        (new PartidaService()).cierrePartida(partida);
 
         assertEquals(EstadoPartidaEnum.FINALIZADA, partida.getEstadoPartida());
         assertEquals(partida.getMesa().getJugadorList().get(1), partida.getJugadorGanador());
@@ -44,12 +44,12 @@ class PartidaCierrreServiceImplTest {
 
         listJugadores = creadorListaJugardoresPartidaConLaMismaEstrategia(2, (new PrimeraCartaDeLaManoStrategy()));
 
-        Partida partida = (new PartidaServiceImpl()).crearPartida(listJugadores);
+        Partida partida = (new PartidaService()).crearPartida(listJugadores);
 
         partida.getMesa().getJugadorList().get(0).setPuntuacion(70);
         partida.getMesa().getJugadorList().get(1).setPuntuacion(50);
 
-        (new PartidaServiceImpl()).cierrePartida(partida);
+        (new PartidaService()).cierrePartida(partida);
 
         assertEquals(EstadoPartidaEnum.FINALIZADA, partida.getEstadoPartida());
         assertEquals(partida.getMesa().getJugadorList().get(0), partida.getJugadorGanador());
